+++ event_sourcing/lib/src/storage/postgres/postgres_backend.dart
@@ -881,7 +881,7 @@ class PostgresBackend extends StorageBackend {
         'create one.',
       );
     }
-    event.requireRecordTimestamp();
+    event.requireRecordTimestamps();
     final record = event.toMap();
     await session.execute(
       Sql.named('''